return int from wxGetWindowId(), not WORD, to ensure comparing it with wxWindowID elsewhere works correctly even for negative ids
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@50181 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/mdi.cpp
@@ -601,8 +601,8 @@ bool wxMDIParentFrame::HandleCommand(WXWORD id, WXWORD cmd, WXHWND hwnd)
             wxWindow *child = node->GetData();
             if ( child->GetHWND() )
             {
-                long childId = wxGetWindowId(child->GetHWND());
-                if (childId == (long)id)
+                int childId = wxGetWindowId(child->GetHWND());
+                if ( childId == (signed short)id )
                 {
                     ::SendMessage( GetWinHwnd(GetClientWindow()),
                                    WM_MDIACTIVATE,
